requirements.txt
spacy
nltk
numpy
gensim
sentence_transformers
torch
transformers
scikit-learn
google-generativeai